공지사항 목록, 상세보기, 삭제 구현

+++ src/main/java/com/plum/notice/dto/NoticeDTO.java
@@ -8,7 +8,7 @@ import lombok.ToString;
 @Setter
 @ToString
 public class NoticeDTO {
-    private int no;
+    private long no;
     private String title;
     private String content;
     private String user;
